+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 Release 5.6.0:
  - OpenID for Verifiable Credential Issuance:
    - Remove code elements deprecated in 5.5.0
+ - OpenID for Verifiable Presentations:
+   - In `OpenId4VpVerifier` add constructor parameter `supportedAlgorithms`
+ - Use function decomposition:
+   - Replace `VerifierCryptoService` with `VerifySignatureFun`
 
 Release 5.5.0:
  - Remove elements deprecated in 5.4.0 when introducing DCQL: